backend/requirements.txt
flask
flask-cors
tensorflow==2.15.0
numpy
pillow
scikit-learn==1.6.1
gunicorn